minishell with my implementation of the following commands: cd, pwd, echo, export, unset, env, exit

+++ srcs/standart_split.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   standart_split.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rnancee <rnancee@student.42.fr>            +#+  +:+       +#+        */
+/*   By: rmass <rmass@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/23 17:28:20 by rmass             #+#    #+#             */
-/*   Updated: 2021/02/23 17:30:30 by rnancee          ###   ########.fr       */
+/*   Updated: 2021/03/20 20:26:17 by rmass            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 
 	if (s == 0)
 		return (0);
-	if (ft_strlen(s) <= start)
+	if (ft_strlen(s) <= (int)start)
 		return (ft_strdup(""));
 	if (!(res = malloc(len + 1)))
 		return (0);
